requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
av==14.3.0
boto3==1.37.32
botocore==1.37.32
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.12
googleapis-common-protos==1.59.1
grpcio==1.71.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
jmespath==1.0.1
numpy==2.2.4
opencv-python-headless==4.11.0.86
opentelemetry-api==1.32.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
packaging==24.2
protobuf==4.25.6
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
thrift==0.21.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0